fix(chat): harden Doc Query proof receipts
